Fix zkSync payer panic when no receipt is found (#7)
If the receipt is not found, DefaultProvider.GetTransactionReceipt
returns nil. The code does not handle this case and panics checking for
the receipt status.

This change adjusts the code to consider the status "pending" if no
receipt is found (e.g. receipt is nil), "failed" if the receipt is found
with a non-successful status, and "success" otherwise.